+++ src/router.ts
@@ -4,6 +4,7 @@ import { routeTree } from './routeTree.gen';
 
 export const router = createRouter({
   routeTree,
+  scrollToTopSelectors: ["[data-main-scroll]"],
   context: {
     queryClient,
     auth: {